refactor location of primitives
