+++ include/NeoFOAM/core/primitives/scalar.hpp
