+++ test/fields/test_DomainField.cpp
@@ -8,6 +8,7 @@
 
 #include "NeoFOAM/fields/BoundaryFields.hpp"
 #include "NeoFOAM/fields/DomainField.hpp"
+#include "NeoFOAM/fields/comparisons/FieldComparison.hpp"
 
 TEST_CASE("Boundaries")
 {
